pyproject.toml
[tool.poetry]
name = "wordle_solver"
version = "0.1.0"
description = ""
authors = ["Yuki Shiga <y.shiga.91+yush1ga@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
nltk = "^3.6.7"
tqdm = "^4.62.3"
python-Levenshtein = "^0.12.2"

[tool.poetry.dev-dependencies]
mypy = "^0.931"
black = {version = "^21.12b0", allow-prereleases = true}
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
